refactor: projects -> sources.projects, components -> sources.components
